pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-action-policy"
version = "0.1.0"
description = "Declarative action policies for AI agents — approve, deny, or escalate any tool call before execution"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "hmj", email = "hmj@quartzunit.com" }]
keywords = [
    "agent",
    "policy",
    "safety",
    "guardrail",
    "tool-use",
    "llm",
    "ai-agent",
    "security",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Typing :: Typed",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/QuartzUnit/agent-action-policy"
Repository = "https://github.com/QuartzUnit/agent-action-policy"
Issues = "https://github.com/QuartzUnit/agent-action-policy/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/action_policy"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
